# Users/api/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API to manage users.",
        "title": "middleware/api/user",
        "contact": {
            "name": "Justine Bachelard.",
            "email": "justine.bachelard@ext.uca.fr"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
        "/user": {
            "get": {
                "description": "Get users.",
                "tags": [
                    "users"
                ],
                "summary": "Get users.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    },
                    "500": {
                        "description": "Something went wrong"
                    }
                }
            },
            "post": {
                "description": "Adding a user.",
                "tags": [
                    "postUser"
                ],
                "summary": "Post user.",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Something went wrong"
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "description": "Get a user.",
                "tags": [
                    "users"
                ],
                "summary": "Get a user.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User UUID formatted ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "422": {
                        "description": "Cannot parse id"
                    },
                    "500": {
                        "description": "Something went wrong"
                    }
                }
            },
            "put": {
                "description": "update a user.",
                "tags": [
                    "updateUsers"
                ],
                "summary": "update a user.",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Something went wrong"
                    }
                }
            },
            "delete": {
                "description": "Delete user.",
                "tags": [
                    "deleteUser"
                ],
                "summary": "Delete user.",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Something went wrong"
                    }
                }
            }
        }
    },
    "definitions": {
        "models.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "mail": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}